# attendance-backend-main/public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"ServiceResponse_any_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"result": {
						"nullable": true
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success",
					"message",
					"result",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": true
			},
			"CheckInRequest": {
				"properties": {
					"visitor": {
						"properties": {
							"idNumber": {
								"type": "string"
							},
							"idProofType": {
								"type": "string"
							},
							"hostName": {
								"type": "string"
							},
							"department": {
								"type": "string"
							},
							"purpose": {
								"type": "string"
							},
							"visitorCompany": {
								"type": "string"
							},
							"email": {
								"type": "string"
							},
							"mobile": {
								"type": "string"
							},
							"fullName": {
								"type": "string"
							}
						},
						"required": [
							"fullName"
						],
						"type": "object"
					},
					"verificationMode": {
						"type": "string"
					},
					"appointmentId": {
						"type": "string"
					},
					"department": {
						"type": "string"
					},
					"badgeType": {
						"type": "string",
						"enum": [
							"QR",
							"RFID"
						]
					}
				},
				"required": [
					"visitor"
				],
				"type": "object",
				"additionalProperties": {}
			},
...
			"AttendanceResponse": {
				"properties": {
					"id": {
						"type": "string"
					},
					"userId": {
						"type": "string"
					},
					"checkIn": {
						"type": "string",
						"format": "date-time"
					},
					"checkOut": {
						"type": "string",
						"format": "date-time"
					},
					"hoster": {
						"type": "string"
					},
					"badge": {
						"type": "string"
					},
					"date": {
						"type": "string"
					},
					"status": {
						"type": "string"
					},
					"note": {
						"type": "string"
					},
					"user": {},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"userId",
					"checkIn",
					"date",
					"status"
				],
				"type": "object",
				"additionalProperties": true
			},
			"AttendanceListResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"result": {
						"items": {
							"$ref": "#/components/schemas/AttendanceResponse"
						},
						"type": "array",
						"nullable": true
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success",
					"message",
					"result",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": true
			},
			"AttendanceSingleResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"result": {
						"allOf": [
							{
								"$ref": "#/components/schemas/AttendanceResponse"
							}
						],
						"nullable": true
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success",
					"message",
					"result",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": true
			},
			"CheckOutRequest": {
				"properties": {
					"note": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": true
			},
			"UpdateAttendanceRequest": {
				"properties": {
					"status": {
						"type": "string",
						"enum": [
							"present",
							"late",
							"left_early",
							"absent"
						]
					},
					"note": {
						"type": "string"
					},
					"checkOut": {
						"type": "string",
						"format": "date-time"
					},
					"checkIn": {
						"type": "string",
						"format": "date-time"
					}
				},
				"type": "object",
				"additionalProperties": true
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header",
				"description": "JWT token (Bearer <token>)"
			}
		}
	},
	"info": {
		"title": "Attendance Backend API",
		"version": "1.0.0",
		"description": "API documentation for attendance management system",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "API Support"
		}
	},
	"paths": {
...
		"/api/visitors/check-in": {
			"post": {
				"operationId": "CheckIn",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_any_"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_any_"
								}
							}
						}
					},
					"403": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_any_"
								}
							}
						}
					}
				},
				"tags": [
					"Visitors"
				],
				"security": [
					{
						"jwt": [
							"visitor:create"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CheckInRequest"
							}
						}
					}
				}
			}
		},
		"/api/visitors/check-out": {
			"post": {
				"operationId": "CheckOut",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_any_"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_any_"
								}
							}
						}
					}
				},
				"tags": [
					"Visitors"
				],
				"security": [
					{
						"jwt": [
							"visitor:update"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"badgeId": {
										"type": "string"
									},
									"visitorId": {
										"type": "string"
									}
								},
								"type": "object"
							}
						}
					}
				}
			}
		},
...
		"/api/attendance": {
			"get": {
				"operationId": "GetAllAttendance",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AttendanceListResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Attendance"
				],
				"security": [
					{
						"jwt": [
							"attendance:read"
						]
					}
				],
				"parameters": []
			}
		},
		"/api/attendance/{id}": {
			"get": {
				"operationId": "GetAttendanceById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AttendanceSingleResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Attendance"
				],
				"security": [
					{
						"jwt": [
							"attendance:read"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateAttendance",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AttendanceSingleResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Attendance"
				],
				"security": [
					{
						"jwt": [
							"attendance:update"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateAttendanceRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteAttendance",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AttendanceSingleResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Attendance"
				],
				"security": [
					{
						"jwt": [
							"attendance:delete"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/attendance/by-user/{userId}": {
			"get": {
				"operationId": "GetAttendanceByUserId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AttendanceListResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Attendance"
				],
				"security": [
					{
						"jwt": [
							"attendance:read"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/attendance/by-date/{date}": {
			"get": {
				"operationId": "GetAttendanceByDate",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AttendanceListResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Attendance"
				],
				"security": [
					{
						"jwt": [
							"attendance:read"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "date",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/attendance/report/range/{startDate}/{endDate}": {
			"get": {
				"operationId": "GetAttendanceReport",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AttendanceListResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Attendance"
				],
				"security": [
					{
						"jwt": [
							"attendance:read"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "startDate",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "endDate",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/attendance/report/user/{userId}/{startDate}/{endDate}": {
			"get": {
				"operationId": "GetUserAttendanceReport",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AttendanceListResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Attendance"
				],
				"security": [
					{
						"jwt": [
							"attendance:read"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "startDate",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "endDate",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/attendance/{id}/checkout": {
			"put": {
				"operationId": "CheckOut",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AttendanceSingleResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Attendance"
				],
				"security": [
					{
						"jwt": [
							"attendance:update"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CheckOutRequest"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}